+++ pom.xml
@@ -17,7 +17,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <dropwizard.version>2.0.3</dropwizard.version>
-        <mainClass>org.example.TestApplication</mainClass>
+        <mainClass>org.soniakbew.TestApplication</mainClass>
     </properties>
 
     <dependencyManagement>
