+++ src/main/java/org/soniakbew/TestConfiguration.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.soniakbew;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.dropwizard.Configuration;
