+++ src/main/java/org/soniakbew/controllers/TestController.java
@@ -1,7 +1,7 @@
-package org.example.controllers;
+package org.soniakbew.controllers;
 
 import io.swagger.annotations.Api;
-import org.example.services.TestService;
+import org.soniakbew.services.TestService;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
