+++ src/main/java/org/soniakbew/daos/DatabaseConnector.java
@@ -1,4 +1,4 @@
-package org.example.daos;
+package org.soniakbew.daos;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
