+++ src/main/java/org/soniakbew/daos/TestDao.java
@@ -1,4 +1,4 @@
-package org.example.daos;
+package org.soniakbew.daos;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
